pyproject.toml
[project]
name = "meeting-intelligance-systems"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.128.0",
    "openai>=2.16.0",
    "pydantic>=2.12.5",
    "python-dotenv>=1.2.1",
    "python-multipart>=0.0.22",
    "pyyaml>=6.0",
    "qdrant-client>=1.16.2",
    "uvicorn>=0.40.0",
    "streamlit>=1.31.0",
    "requests>=2.31.0",
]

[dependency-groups]
dev = [
    "httpx>=0.28.1",
    "pyright>=1.1.408",
    "pytest>=9.0.2",
    "pytest-html>=4.2.0",
    "ruff>=0.14.14",
]
